Added concepts header to pch.
Added helper functions to common to get current and previous dates.

Added a type to handle file parsing. Added a builder for that type to make it easy to decide which headers to look for when printing log info.

Cleaned up functions in main. Main now uses a temporary global parser variable to parse the log files.